requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
defusedxml==0.7.1
Django==3.2
django-allauth==0.50.0
django-environ==0.8.1
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
idna==3.3
mysqlclient==2.1.0
oauthlib==3.2.0
pycparser==2.21
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
